+++ MyLearn/MyLearn/InputModels/NewJobComment.cs
@@ -5,6 +5,6 @@
         public string JobOfferId { get; set; }
         public string JobOfferComment { get; set; }
         public string ParentId { get; set; }
-        public int Commenter { get; set; }
+        public float Commenter { get; set; }
     }
 }
