+++ MyLearn/MyLearn/Models/JobOfferBid.cs
@@ -3,7 +3,7 @@
     public class JobOfferBid
     {
         public string Money { get; set; }
-        public int DurationInDays { get; set; }
+        public float DurationInDays { get; set; }
         public string StudentName { get; set; }
         public string StudentSurname { get; set; }
         public string StudentUserId { get; set; }
